+++ src/test/java/com/attica/athens/domain/chat/domain/ChatsTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.BDDAssertions.then;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
 
+import com.attica.athens.support.TestDoubleFactory;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
